refactor: Refactor Reservation Modal
